rename modules and move to new structure
