+++ src/app/modules/home/components/create-share-section/register.component.scss
