+++ src/app/modules/home/components/explore-section/explore-section.component.html
