+++ src/app/modules/home/components/home-section-registerForm/homes-section-registerForm.component.scss
@@ -3,7 +3,7 @@
    }
   
   #home-section {
-    background: url(../../../../../../assets/AloneOak.jpg);
+    background: url(../../../../../assets/AloneOak.jpg);
     background-repeat: no-repeat;
     background-size: cover;
     background-attachment: fixed;
