+++ src/app/modules/home/home.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Routes, RouterModule } from '@angular/router';
-import { RegisterComponent } from './register/register.component';
-import { HomeSectionRegisterFormComponent } from './register/components/home-section-registerForm/homes-section-registerForm.component';
-import { NavComponent } from './register/components/nav/nav.component';
-import { ExploreSectionComponent } from './register/components/explore-section/explore-section.component';
-import { CreateShareSectionComponent } from './register/components/create-share-section/create-share-section.component';
-import { FooterSectionComponent } from './register/components/footer-section/footer-section.component';
+import { HomeComponent } from './components/home.component';
+import { HomeSectionRegisterFormComponent } from './components/home-section-registerForm/homes-section-registerForm.component';
+import { NavComponent } from './components/nav/nav.component';
+import { ExploreSectionComponent } from './components/explore-section/explore-section.component';
+import { CreateShareSectionComponent } from './components/create-share-section/create-share-section.component';
+import { FooterSectionComponent } from './components/footer-section/footer-section.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RegisterFormComponent } from 'src/app/components/register-form/register-form.component';
 import { SnacbarAlertService } from '../../services/snacbar-alert.service';
@@ -14,13 +14,13 @@ import { NgxSpinnerModule } from 'ngx-spinner';
 import { MaterialModule } from '../shared/material/material.module';
 
 const routes: Routes = [
-  {path: '', component: RegisterComponent},
+  {path: '', component: HomeComponent},
   {path: '**', redirectTo: ''}
 ];
 
 @NgModule({
   declarations: [
-    RegisterComponent,
+    HomeComponent,
     HomeSectionRegisterFormComponent,
     NavComponent,
     ExploreSectionComponent,
@@ -40,4 +40,4 @@ const routes: Routes = [
     SnacbarAlertService
   ]
 })
-export class HomeRegisterModule { }
+export class HomeModule { }
